Force upgrade for "pip.installed" states
Without this, salt would (errouneoulsy) produce an empty packages
argument for "pip install" call, thus resulting in a error saying "You
must give at least one requirement to install".

Perhaps this is fixed in more recent versions of Salt? I didn't checked.

In the meantime, we force upgrade so that idempotence works.
Accordingly, we add pip.installed state to the list of "exclude" in
test_idempotence so that their "Changes" result won't be inspected.